del: ab 테스트 종료로 인한 homeViewControllerB 없앰
